move another enum
